backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-decouple
djangorestframework-simplejwt
requests
